+++ shared/protocol/build.rs
@@ -26,6 +26,7 @@ fn compile_protos() {
                 "proto/reovim/v2/syntax.proto",
                 "proto/reovim/v2/presence.proto",
                 "proto/reovim/v2/debug.proto",
+                "proto/reovim/v2/extension.proto",
             ],
             &["proto"],
         )